+++ src/fountain.c
@@ -1014,7 +1014,7 @@ register struct obj *obj;
 					money = somegold(money) / 10;
 					for (otmp = invent; otmp && money > 0; otmp = otmp->nobj)
 						if (otmp->oclass == COIN_CLASS) {
-							int denomination = objects[otmp->otyp].oc_cost;
+							long denomination = objects[otmp->otyp].oc_cost;
 							long coin_loss =
 								(money + denomination - 1) / denomination;
 							coin_loss = min(coin_loss, otmp->quan);
